--- src/WebApiClient/Digdir.Library.Dialogporten.WebApiClient.EndUser/Features/V1/AcceptedLanguagesHeaderFormat.cs
@@ -1,17 +1,0 @@
-using Altinn.ApiClients.Dialogporten.Common;
-
-namespace Altinn.ApiClients.Dialogporten.EndUser.Features.V1;
-
-public partial class AcceptedLanguages
-{
-    public override string ToString() =>
-        AcceptedLanguagesHeaderFormatter.FormatAcceptedLanguages(
-            AcceptedLanguage,
-            static language => language.ToString());
-}
-
-public partial class AcceptedLanguage
-{
-    public override string ToString() =>
-        AcceptedLanguagesHeaderFormatter.FormatAcceptedLanguage(LanguageCode, Weight);
-}
