+++ src/WebApiClient/Digdir.Library.Dialogporten.WebApiClient.EndUser/Features/V1/Contracts/Activity/DialogActivitySearchItem.cs
@@ -1,0 +1,25 @@
+using System.Text.Json.Serialization;
+
+namespace Altinn.ApiClients.Dialogporten.EndUser.Features.V1.Contracts.Activity;
+
+public class DialogActivitySearchItem
+{
+    [JsonPropertyName("id")]
+    public Guid Id { get; set; }
+
+    [JsonPropertyName("createdAt")]
+    public DateTimeOffset CreatedAt { get; set; }
+
+    [JsonPropertyName("extendedType")]
+    public Uri? ExtendedType { get; set; }
+
+    [JsonPropertyName("type")]
+    [JsonConverter(typeof(JsonStringEnumConverter<DialogActivityType>))]
+    public DialogActivityType Type { get; set; }
+
+    [JsonPropertyName("transmissionId")]
+    public Guid? TransmissionId { get; set; }
+
+    [JsonPropertyName("description")]
+    public ICollection<Localization>? Description { get; set; }
+}
