+++ src/WebApiClient/Digdir.Library.Dialogporten.WebApiClient.EndUser/Features/V1/Contracts/Contents/Localization.cs
@@ -1,0 +1,18 @@
+using System.Text.Json.Serialization;
+
+namespace Altinn.ApiClients.Dialogporten.EndUser.Features.V1.Contracts.Contents;
+
+public class Localization
+{
+    /// <summary>
+    /// The localized text (or URL if a front-channel embed).
+    /// </summary>
+    [JsonPropertyName("value")]
+    public string Value { get; set; } = null!;
+
+    /// <summary>
+    /// The language code of the localization in ISO 639-1 format.
+    /// </summary>
+    [JsonPropertyName("languageCode")]
+    public string LanguageCode { get; set; } = null!;
+}
