+++ src/WebApiClient/Digdir.Library.Dialogporten.WebApiClient.EndUser/Features/V1/Contracts/GlobalUsings.cs
@@ -1,0 +1,13 @@
+global using Altinn.ApiClients.Dialogporten.EndUser.Features.V1.Contracts.Action;
+global using Altinn.ApiClients.Dialogporten.EndUser.Features.V1.Contracts.Activity;
+global using Altinn.ApiClients.Dialogporten.EndUser.Features.V1.Contracts.Actors;
+global using Altinn.ApiClients.Dialogporten.EndUser.Features.V1.Contracts.Attachment;
+global using Altinn.ApiClients.Dialogporten.EndUser.Features.V1.Contracts.Common;
+global using Altinn.ApiClients.Dialogporten.EndUser.Features.V1.Contracts.Contents;
+global using Altinn.ApiClients.Dialogporten.EndUser.Features.V1.Contracts.Dialog;
+global using Altinn.ApiClients.Dialogporten.EndUser.Features.V1.Contracts.IdentifierLookup;
+global using Altinn.ApiClients.Dialogporten.EndUser.Features.V1.Contracts.Labels;
+global using Altinn.ApiClients.Dialogporten.EndUser.Features.V1.Contracts.Party;
+global using Altinn.ApiClients.Dialogporten.EndUser.Features.V1.Contracts.Search;
+global using Altinn.ApiClients.Dialogporten.EndUser.Features.V1.Contracts.ServiceResource;
+global using Altinn.ApiClients.Dialogporten.EndUser.Features.V1.Contracts.Transmission;
