+++ src/WebApiClient/Digdir.Library.Dialogporten.WebApiClient.EndUser/Features/V1/IEnduserApi.cs
@@ -1,0 +1,613 @@
+using Altinn.ApiClients.Dialogporten.EndUser.Features.V1.Contracts;
+using Altinn.ApiClients.Maskinporten.Models;
+using Refit;
+
+namespace Altinn.ApiClients.Dialogporten.EndUser.Features.V1;
+
+/// <summary>Gets the service resources the authenticated end user is authorized to use.</summary>
+public interface IEnduserApi
+{
+    /// <summary>Gets the service resources the authenticated end user is authorized to use.</summary>
+    /// <remarks>Returns the same service resource metadata as the public metadata endpoint, filtered to the resources the calling end user is authorized to use. Optionally narrowed by one or more party URNs.</remarks>
+    /// <param name="party">Filter by one or more party URNs. Parties the caller is not authorized for are silently ignored.</param>
+    /// <param name="acceptLanguage">accept_Language parameter</param>
+    /// <param name="cancellationToken">The cancellation token to cancel the request.</param>
+    /// <returns>
+    /// A <see cref="Task"/> representing the <see cref="IApiResponse"/> instance containing the result:
+    /// <list type="table">
+    /// <listheader>
+    /// <term>Status</term>
+    /// <description>Description</description>
+    /// </listheader>
+    /// <item>
+    /// <term>200</term>
+    /// <description>Authorized service resource metadata.</description>
+    /// </item>
+    /// <item>
+    /// <term>401</term>
+    /// <description>Unauthorized</description>
+    /// </item>
+    /// <item>
+    /// <term>403</term>
+    /// <description>Forbidden</description>
+    /// </item>
+    /// <item>
+    /// <term>503</term>
+    /// <description>Service Unavailable, used when Dialogporten is in maintenance mode</description>
+    /// </item>
+    /// </list>
+    /// </returns>
+    [Headers("Accept: application/json, text/plain")]
+    [Get("/api/v1/enduser/serviceresources")]
+    Task<IApiResponse<ServiceResourceMetadataList>> SearchAuthorizedServiceResources(
+        [Query(CollectionFormat.Multi)] IEnumerable<string>? party = null, [Header("accept-Language")] AcceptedLanguages? acceptLanguage = null,
+        [Property("Altinn.ApiClients.Maskinporten.RequestContext")] MaskinportenRequestContext? requestContext = null, CancellationToken cancellationToken = default);
+
+    /// <summary>Gets a list of dialog label assignment logs</summary>
+    /// <remarks>Gets the list of label assignment logs belonging to a dialog</remarks>
+    /// <param name="dialogId">dialogId parameter</param>
+    /// <param name="cancellationToken">The cancellation token to cancel the request.</param>
+    /// <returns>
+    /// A <see cref="Task"/> representing the <see cref="IApiResponse"/> instance containing the result:
+    /// <list type="table">
+    /// <listheader>
+    /// <term>Status</term>
+    /// <description>Description</description>
+    /// </listheader>
+    /// <item>
+    /// <term>200</term>
+    /// <description>Successfully returned the dialog label assignment log list.</description>
+    /// </item>
+    /// <item>
+    /// <term>401</term>
+    /// <description>Missing or invalid authentication token. Requires a Maskinporten-token with the scope \"digdir:dialogporten\".</description>
+    /// </item>
+    /// <item>
+    /// <term>403</term>
+    /// <description>Unauthorized to get the supplied dialog (not owned by authenticated organization or has additional scope requirements defined in policy).</description>
+    /// </item>
+    /// <item>
+    /// <term>404</term>
+    /// <description>The given dialog ID was not found.</description>
+    /// </item>
+    /// <item>
+    /// <term>410</term>
+    /// <description>Entity with the given key(s) is removed.</description>
+    /// </item>
+    /// <item>
+    /// <term>503</term>
+    /// <description>Service Unavailable, used when Dialogporten is in maintenance mode</description>
+    /// </item>
+    /// </list>
+    /// </returns>
+    [Headers("Accept: application/json, application/problem+json, text/plain")]
+    [Get("/api/v1/enduser/dialogs/{dialogId}/context/labellog")]
+    Task<IApiResponse<ICollection<LabelAssignmentLog>>> SearchDialogLabelAssignmentLogs(
+        Guid dialogId, [Property("Altinn.ApiClients.Maskinporten.RequestContext")] MaskinportenRequestContext? requestContext = null, CancellationToken cancellationToken = default);
+
+    /// <summary>Sets the system labels of a dialog</summary>
+    /// <remarks>
+    /// Sets the system labels of the dialog.
+    /// 
+    /// Optimistic concurrency control is implemented using the If-Match header. Supply EnduserContextRevision to ensure that the context is not modified/deleted by another request in the meantime.
+    /// </remarks>
+    /// <param name="dialogId">dialogId parameter</param>
+    /// <param name="ifMatch">if_Match parameter</param>
+    /// <param name="setDialogSystemLabelRequest">setDialogSystemLabelRequest parameter</param>
+    /// <param name="cancellationToken">The cancellation token to cancel the request.</param>
+    /// <returns>
+    /// A <see cref="Task"/> representing the <see cref="IApiResponse"/> instance containing the result:
+    /// <list type="table">
+    /// <listheader>
+    /// <term>Status</term>
+    /// <description>Description</description>
+    /// </listheader>
+    /// <item>
+    /// <term>204</term>
+    /// <description>The dialog system label was updated successfully.</description>
+    /// </item>
+    /// <item>
+    /// <term>400</term>
+    /// <description>Validation error occurred. See problem details for a list of errors.</description>
+    /// </item>
+    /// <item>
+    /// <term>401</term>
+    /// <description>Missing or invalid authentication token. Requires a Maskinporten-token with the scope \"digdir:dialogporten\".</description>
+    /// </item>
+    /// <item>
+    /// <term>403</term>
+    /// <description>Unauthorized to update the supplied dialog (not owned by authenticated organization or has additional scope requirements defined in policy).</description>
+    /// </item>
+    /// <item>
+    /// <term>404</term>
+    /// <description>The given dialog ID was not found.</description>
+    /// </item>
+    /// <item>
+    /// <term>409</term>
+    /// <description>Conflict occurred while processing the request.</description>
+    /// </item>
+    /// <item>
+    /// <term>410</term>
+    /// <description>Entity with the given key(s) is removed.</description>
+    /// </item>
+    /// <item>
+    /// <term>412</term>
+    /// <description>The supplied If-Match header did not match the current Revision value for the dialog. The request was not applied.</description>
+    /// </item>
+    /// <item>
+    /// <term>422</term>
+    /// <description>Domain error occurred. See problem details for a list of errors.</description>
+    /// </item>
+    /// <item>
+    /// <term>503</term>
+    /// <description>Service Unavailable, used when Dialogporten is in maintenance mode</description>
+    /// </item>
+    /// </list>
+    /// </returns>
+    [Headers("Accept: application/problem+json, text/plain", "Content-Type: application/json")]
+    [Put("/api/v1/enduser/dialogs/{dialogId}/context/systemlabels")]
+    Task<IApiResponse> SetDialogSystemLabels(
+        Guid dialogId, [Body, AliasAs("SetDialogSystemLabelRequest")] SetDialogSystemLabelRequest setDialogSystemLabelRequest, [Header("if-Match")] Guid? ifMatch = null,
+        [Property("Altinn.ApiClients.Maskinporten.RequestContext")] MaskinportenRequestContext? requestContext = null, CancellationToken cancellationToken = default);
+
+    /// <summary>Sets system labels for multiple dialogs</summary>
+    /// <remarks>Sets the system labels for a list of dialogs, optionally including a end user context revision for each dialog.</remarks>
+    /// <param name="dto">dto parameter</param>
+    /// <param name="cancellationToken">The cancellation token to cancel the request.</param>
+    /// <returns>
+    /// A <see cref="Task"/> representing the <see cref="IApiResponse"/> instance containing the result:
+    /// <list type="table">
+    /// <listheader>
+    /// <term>Status</term>
+    /// <description>Description</description>
+    /// </listheader>
+    /// <item>
+    /// <term>204</term>
+    /// <description>The dialog system labels was updated successfully.</description>
+    /// </item>
+    /// <item>
+    /// <term>400</term>
+    /// <description>Validation error occurred. See problem details for a list of errors.</description>
+    /// </item>
+    /// <item>
+    /// <term>401</term>
+    /// <description>Missing or invalid authentication token. Requires a Maskinporten-token with the scope \"digdir:dialogporten\".</description>
+    /// </item>
+    /// <item>
+    /// <term>403</term>
+    /// <description>Unauthorized to update the supplied dialog (not owned by authenticated organization or has additional scope requirements defined in policy).</description>
+    /// </item>
+    /// <item>
+    /// <term>409</term>
+    /// <description>Conflict occurred while processing the request.</description>
+    /// </item>
+    /// <item>
+    /// <term>412</term>
+    /// <description>The supplied If-Match header did not match the current Revision value for the dialog. The request was not applied.</description>
+    /// </item>
+    /// <item>
+    /// <term>422</term>
+    /// <description>Domain error occurred. See problem details for a list of errors.</description>
+    /// </item>
+    /// <item>
+    /// <term>503</term>
+    /// <description>Service Unavailable, used when Dialogporten is in maintenance mode</description>
+    /// </item>
+    /// </list>
+    /// </returns>
+    [Headers("Accept: application/problem+json, text/plain", "Content-Type: application/json")]
+    [Post("/api/v1/enduser/dialogs/context/systemlabels/actions/bulkset")]
+    Task<IApiResponse> BulkSetDialogSystemLabels(
+        [Body] BulkSetSystemLabel dto, [Property("Altinn.ApiClients.Maskinporten.RequestContext")] MaskinportenRequestContext? requestContext = null, CancellationToken cancellationToken = default);
+
+    /// <summary>Gets a list of dialog transmissions</summary>
+    /// <remarks>Gets the list of transmissions belonging to a dialog</remarks>
+    /// <param name="dialogId">dialogId parameter</param>
+    /// <param name="acceptLanguage">accept_Language parameter</param>
+    /// <param name="cancellationToken">The cancellation token to cancel the request.</param>
+    /// <returns>
+    /// A <see cref="Task"/> representing the <see cref="IApiResponse"/> instance containing the result:
+    /// <list type="table">
+    /// <listheader>
+    /// <term>Status</term>
+    /// <description>Description</description>
+    /// </listheader>
+    /// <item>
+    /// <term>200</term>
+    /// <description>Successfully returned the dialog transmission list.</description>
+    /// </item>
+    /// <item>
+    /// <term>401</term>
+    /// <description>Missing or invalid authentication token. Requires a Maskinporten-token with the scope \"digdir:dialogporten\".</description>
+    /// </item>
+    /// <item>
+    /// <term>403</term>
+    /// <description>Unauthorized to get the supplied dialog (not owned by authenticated organization or has additional scope requirements defined in policy).</description>
+    /// </item>
+    /// <item>
+    /// <term>404</term>
+    /// <description>The given dialog ID was not found.</description>
+    /// </item>
+    /// <item>
+    /// <term>410</term>
+    /// <description>Entity with the given key(s) is removed.</description>
+    /// </item>
+    /// <item>
+    /// <term>503</term>
+    /// <description>Service Unavailable, used when Dialogporten is in maintenance mode</description>
+    /// </item>
+    /// </list>
+    /// </returns>
+    [Headers("Accept: application/json, application/problem+json, text/plain")]
+    [Get("/api/v1/enduser/dialogs/{dialogId}/transmissions")]
+    Task<IApiResponse<ICollection<DialogTransmissionDetails>>> SearchDialogTransmissions(
+        Guid dialogId, [Header("accept-Language")] AcceptedLanguages? acceptLanguage = null,
+        [Property("Altinn.ApiClients.Maskinporten.RequestContext")] MaskinportenRequestContext? requestContext = null, CancellationToken cancellationToken = default);
+
+    /// <summary>Gets all seen log records for a dialog</summary>
+    /// <remarks>Gets all seen log records for a dialog.</remarks>
+    /// <param name="dialogId">dialogId parameter</param>
+    /// <param name="cancellationToken">The cancellation token to cancel the request.</param>
+    /// <returns>
+    /// A <see cref="Task"/> representing the <see cref="IApiResponse"/> instance containing the result:
+    /// <list type="table">
+    /// <listheader>
+    /// <term>Status</term>
+    /// <description>Description</description>
+    /// </listheader>
+    /// <item>
+    /// <term>200</term>
+    /// <description>Successfully returned the dialog seen log records.</description>
+    /// </item>
+    /// <item>
+    /// <term>401</term>
+    /// <description>Missing or invalid authentication token. Requires a Maskinporten-token with the scope \"digdir:dialogporten\".</description>
+    /// </item>
+    /// <item>
+    /// <term>403</term>
+    /// <description>Forbidden</description>
+    /// </item>
+    /// <item>
+    /// <term>404</term>
+    /// <description>Not Found</description>
+    /// </item>
+    /// <item>
+    /// <term>503</term>
+    /// <description>Service Unavailable, used when Dialogporten is in maintenance mode</description>
+    /// </item>
+    /// </list>
+    /// </returns>
+    [Headers("Accept: application/json, application/problem+json, text/plain")]
+    [Get("/api/v1/enduser/dialogs/{dialogId}/seenlog")]
+    Task<IApiResponse<ICollection<DialogSeenLogDetails>>> SearchDialogSeenLogs(
+        Guid dialogId, [Property("Altinn.ApiClients.Maskinporten.RequestContext")] MaskinportenRequestContext? requestContext = null, CancellationToken cancellationToken = default);
+
+    /// <summary>Gets a list of dialog activities</summary>
+    /// <remarks>Gets the list of activities belonging to a dialog</remarks>
+    /// <param name="dialogId">dialogId parameter</param>
+    /// <param name="acceptLanguage">accept_Language parameter</param>
+    /// <param name="cancellationToken">The cancellation token to cancel the request.</param>
+    /// <returns>
+    /// A <see cref="Task"/> representing the <see cref="IApiResponse"/> instance containing the result:
+    /// <list type="table">
+    /// <listheader>
+    /// <term>Status</term>
+    /// <description>Description</description>
+    /// </listheader>
+    /// <item>
+    /// <term>200</term>
+    /// <description>Successfully returned the dialog activity list.</description>
+    /// </item>
+    /// <item>
+    /// <term>401</term>
+    /// <description>Missing or invalid authentication token. Requires a Maskinporten-token with the scope \"digdir:dialogporten\".</description>
+    /// </item>
+    /// <item>
+    /// <term>403</term>
+    /// <description>Unauthorized to get the supplied dialog (not owned by authenticated organization or has additional scope requirements defined in policy).</description>
+    /// </item>
+    /// <item>
+    /// <term>404</term>
+    /// <description>The given dialog ID was not found.</description>
+    /// </item>
+    /// <item>
+    /// <term>503</term>
+    /// <description>Service Unavailable, used when Dialogporten is in maintenance mode</description>
+    /// </item>
+    /// </list>
+    /// </returns>
+    [Headers("Accept: application/json, application/problem+json, text/plain")]
+    [Get("/api/v1/enduser/dialogs/{dialogId}/activities")]
+    Task<IApiResponse<ICollection<DialogActivitySearchItem>>> SearchDialogActivities(
+        Guid dialogId, [Header("accept-Language")] AcceptedLanguages? acceptLanguage = null,
+        [Property("Altinn.ApiClients.Maskinporten.RequestContext")] MaskinportenRequestContext? requestContext = null, CancellationToken cancellationToken = default);
+
+    /// <summary>Gets a list of dialogs</summary>
+    /// <remarks>
+    /// Performs a search for dialogs, returning a paginated list of dialogs.
+    /// 
+    /// * All date parameters must contain explicit time zone. Example: 2023-10-27T10:00:00Z or 2023-10-27T10:00:00+01:00
+    /// * See "continuationToken" in the response for how to get the next page of results.
+    /// * hasNextPage will be set to true if there are more items to get.
+    /// </remarks>
+    /// <param name="acceptLanguage">accept_Language parameter</param>
+    /// <param name="queryParams">The dynamic querystring parameter wrapping all others.</param>
+    /// <param name="cancellationToken">The cancellation token to cancel the request.</param>
+    /// <returns>
+    /// A <see cref="Task"/> representing the <see cref="IApiResponse"/> instance containing the result:
+    /// <list type="table">
+    /// <listheader>
+    /// <term>Status</term>
+    /// <description>Description</description>
+    /// </listheader>
+    /// <item>
+    /// <term>200</term>
+    /// <description>Successfully returned the dialog list.</description>
+    /// </item>
+    /// <item>
+    /// <term>401</term>
+    /// <description>Missing or invalid authentication token. Requires a Maskinporten-token with the scope \"digdir:dialogporten\".</description>
+    /// </item>
+    /// <item>
+    /// <term>403</term>
+    /// <description>Forbidden</description>
+    /// </item>
+    /// <item>
+    /// <term>422</term>
+    /// <description>Domain error occurred. See problem details for a list of errors.</description>
+    /// </item>
+    /// <item>
+    /// <term>503</term>
+    /// <description>Service Unavailable, used when Dialogporten is in maintenance mode</description>
+    /// </item>
+    /// </list>
+    /// </returns>
+    [Headers("Accept: application/json, application/problem+json, text/plain")]
+    [Get("/api/v1/enduser/dialogs")]
+    Task<IApiResponse<PaginatedDialogSearchItems>> SearchDialogs(
+        [Query] SearchDialogsQueryParams? queryParams = null, [Header("accept-Language")] AcceptedLanguages? acceptLanguage = null,
+        [Property("Altinn.ApiClients.Maskinporten.RequestContext")] MaskinportenRequestContext? requestContext = null, CancellationToken cancellationToken = default);
+
+    /// <summary>Gets a single dialog transmission</summary>
+    /// <remarks>Gets a single transmission belonging to a dialog.</remarks>
+    /// <param name="dialogId">dialogId parameter</param>
+    /// <param name="transmissionId">transmissionId parameter</param>
+    /// <param name="acceptLanguage">accept_Language parameter</param>
+    /// <param name="cancellationToken">The cancellation token to cancel the request.</param>
+    /// <returns>
+    /// A <see cref="Task"/> representing the <see cref="IApiResponse"/> instance containing the result:
+    /// <list type="table">
+    /// <listheader>
+    /// <term>Status</term>
+    /// <description>Description</description>
+    /// </listheader>
+    /// <item>
+    /// <term>200</term>
+    /// <description>Successfully returned the dialog transmission.</description>
+    /// </item>
+    /// <item>
+    /// <term>401</term>
+    /// <description>Missing or invalid authentication token. Requires a Maskinporten-token with the scope \"digdir:dialogporten\".</description>
+    /// </item>
+    /// <item>
+    /// <term>403</term>
+    /// <description>Unauthorized to get child entity for the given dialog (dialog not owned by authenticated organization or has additional scope requirements defined in service identifiers policy).</description>
+    /// </item>
+    /// <item>
+    /// <term>404</term>
+    /// <description>The specified dialog ID or transmission ID was not found.</description>
+    /// </item>
+    /// <item>
+    /// <term>410</term>
+    /// <description>Entity with the given key(s) is removed.</description>
+    /// </item>
+    /// <item>
+    /// <term>503</term>
+    /// <description>Service Unavailable, used when Dialogporten is in maintenance mode</description>
+    /// </item>
+    /// </list>
+    /// </returns>
+    [Headers("Accept: application/json, application/problem+json, text/plain")]
+    [Get("/api/v1/enduser/dialogs/{dialogId}/transmissions/{transmissionId}")]
+    Task<IApiResponse<DialogTransmissionDetails>> GetDialogTransmission(
+        Guid dialogId, Guid transmissionId, [Header("accept-Language")] AcceptedLanguages? acceptLanguage = null,
+        [Property("Altinn.ApiClients.Maskinporten.RequestContext")] MaskinportenRequestContext? requestContext = null, CancellationToken cancellationToken = default);
+
+    /// <summary>Gets a single dialog seen log record</summary>
+    /// <remarks>Gets a single dialog seen log record.</remarks>
+    /// <param name="dialogId">dialogId parameter</param>
+    /// <param name="seenLogId">seenLogId parameter</param>
+    /// <param name="cancellationToken">The cancellation token to cancel the request.</param>
+    /// <returns>
+    /// A <see cref="Task"/> representing the <see cref="IApiResponse"/> instance containing the result:
+    /// <list type="table">
+    /// <listheader>
+    /// <term>Status</term>
+    /// <description>Description</description>
+    /// </listheader>
+    /// <item>
+    /// <term>200</term>
+    /// <description>Successfully returned the dialog seen log record.</description>
+    /// </item>
+    /// <item>
+    /// <term>401</term>
+    /// <description>Missing or invalid authentication token. Requires a Maskinporten-token with the scope \"digdir:dialogporten\".</description>
+    /// </item>
+    /// <item>
+    /// <term>403</term>
+    /// <description>Forbidden</description>
+    /// </item>
+    /// <item>
+    /// <term>404</term>
+    /// <description>Not Found</description>
+    /// </item>
+    /// <item>
+    /// <term>503</term>
+    /// <description>Service Unavailable, used when Dialogporten is in maintenance mode</description>
+    /// </item>
+    /// </list>
+    /// </returns>
+    [Headers("Accept: application/json, application/problem+json, text/plain")]
+    [Get("/api/v1/enduser/dialogs/{dialogId}/seenlog/{seenLogId}")]
+    Task<IApiResponse<DialogSeenLogDetails>> GetDialogSeenLog(
+        Guid dialogId, Guid seenLogId, [Property("Altinn.ApiClients.Maskinporten.RequestContext")] MaskinportenRequestContext? requestContext = null,
+        CancellationToken cancellationToken = default);
+
+    /// <summary>Gets a single dialog activity</summary>
+    /// <remarks>Gets a single activity belonging to a dialog.</remarks>
+    /// <param name="dialogId">dialogId parameter</param>
+    /// <param name="activityId">activityId parameter</param>
+    /// <param name="acceptLanguage">accept_Language parameter</param>
+    /// <param name="cancellationToken">The cancellation token to cancel the request.</param>
+    /// <returns>
+    /// A <see cref="Task"/> representing the <see cref="IApiResponse"/> instance containing the result:
+    /// <list type="table">
+    /// <listheader>
+    /// <term>Status</term>
+    /// <description>Description</description>
+    /// </listheader>
+    /// <item>
+    /// <term>200</term>
+    /// <description>Successfully returned the dialog activity.</description>
+    /// </item>
+    /// <item>
+    /// <term>401</term>
+    /// <description>Missing or invalid authentication token. Requires a Maskinporten-token with the scope \"digdir:dialogporten\".</description>
+    /// </item>
+    /// <item>
+    /// <term>403</term>
+    /// <description>Unauthorized to get child entity for the given dialog (dialog not owned by authenticated organization or has additional scope requirements defined in service identifiers policy).</description>
+    /// </item>
+    /// <item>
+    /// <term>404</term>
+    /// <description>The specified dialog ID or activity ID was not found.</description>
+    /// </item>
+    /// <item>
+    /// <term>503</term>
+    /// <description>Service Unavailable, used when Dialogporten is in maintenance mode</description>
+    /// </item>
+    /// </list>
+    /// </returns>
+    [Headers("Accept: application/json, application/problem+json, text/plain")]
+    [Get("/api/v1/enduser/dialogs/{dialogId}/activities/{activityId}")]
+    Task<IApiResponse<DialogActivity>> GetDialogActivity(
+        Guid dialogId, Guid activityId, [Header("accept-Language")] AcceptedLanguages? acceptLanguage = null,
+        [Property("Altinn.ApiClients.Maskinporten.RequestContext")] MaskinportenRequestContext? requestContext = null, CancellationToken cancellationToken = default);
+
+    /// <summary>Gets a single dialog</summary>
+    /// <remarks>Gets a single dialog aggregate.</remarks>
+    /// <param name="dialogId">dialogId parameter</param>
+    /// <param name="acceptLanguage">accept_Language parameter</param>
+    /// <param name="cancellationToken">The cancellation token to cancel the request.</param>
+    /// <returns>
+    /// A <see cref="Task"/> representing the <see cref="IApiResponse"/> instance containing the result:
+    /// <list type="table">
+    /// <listheader>
+    /// <term>Status</term>
+    /// <description>Description</description>
+    /// </listheader>
+    /// <item>
+    /// <term>200</term>
+    /// <description>Successfully returned the dialog aggregate.</description>
+    /// </item>
+    /// <item>
+    /// <term>401</term>
+    /// <description>Missing or invalid authentication token. Requires a Maskinporten-token with the scope \"digdir:dialogporten\".</description>
+    /// </item>
+    /// <item>
+    /// <term>403</term>
+    /// <description>Unauthorized to get the supplied dialog (not owned by authenticated organization or has additional scope requirements defined in policy).</description>
+    /// </item>
+    /// <item>
+    /// <term>404</term>
+    /// <description>The given dialog ID was not found.</description>
+    /// </item>
+    /// <item>
+    /// <term>503</term>
+    /// <description>Service Unavailable, used when Dialogporten is in maintenance mode</description>
+    /// </item>
+    /// </list>
+    /// </returns>
+    [Headers("Accept: application/json, application/problem+json, text/plain")]
+    [Get("/api/v1/enduser/dialogs/{dialogId}")]
+    Task<IApiResponse<Dialog>> GetDialog(
+        Guid dialogId, [Header("accept-Language")] AcceptedLanguages? acceptLanguage = null,
+        [Property("Altinn.ApiClients.Maskinporten.RequestContext")] MaskinportenRequestContext? requestContext = null, CancellationToken cancellationToken = default);
+
+    /// <summary>Looks up a dialog by instance reference</summary>
+    /// <remarks>Resolves dialog metadata and authorization evidence for a supported instance reference.</remarks>
+    /// <param name="instanceRef">instanceRef parameter</param>
+    /// <param name="acceptLanguage">accept_Language parameter</param>
+    /// <param name="cancellationToken">The cancellation token to cancel the request.</param>
+    /// <returns>
+    /// A <see cref="Task"/> representing the <see cref="IApiResponse"/> instance containing the result:
+    /// <list type="table">
+    /// <listheader>
+    /// <term>Status</term>
+    /// <description>Description</description>
+    /// </listheader>
+    /// <item>
+    /// <term>200</term>
+    /// <description>Successfully resolved dialog lookup metadata.</description>
+    /// </item>
+    /// <item>
+    /// <term>400</term>
+    /// <description>Validation error occurred. See problem details for a list of errors.</description>
+    /// </item>
+    /// <item>
+    /// <term>401</term>
+    /// <description>Missing or invalid authentication token. Requires a Maskinporten-token with the scope \"digdir:dialogporten\".</description>
+    /// </item>
+    /// <item>
+    /// <term>403</term>
+    /// <description>Authenticated end user is not authorized for the supplied instance reference.</description>
+    /// </item>
+    /// <item>
+    /// <term>404</term>
+    /// <description>No dialog match was found for the supplied instance reference.</description>
+    /// </item>
+    /// <item>
+    /// <term>503</term>
+    /// <description>Service Unavailable, used when Dialogporten is in maintenance mode</description>
+    /// </item>
+    /// </list>
+    /// </returns>
+    [Headers("Accept: application/json, application/problem+json, text/plain")]
+    [Get("/api/v1/enduser/dialoglookup")]
+    Task<IApiResponse<EndUserIdentifierLookup>> GetDialogLookup(
+        [Query] string instanceRef, [Header("accept-Language")] AcceptedLanguages? acceptLanguage = null,
+        [Property("Altinn.ApiClients.Maskinporten.RequestContext")] MaskinportenRequestContext? requestContext = null, CancellationToken cancellationToken = default);
+
+    /// <summary>Gets the list of authorized parties for the end user</summary>
+    /// <remarks>Gets the list of authorized parties for the end user.</remarks>
+    /// <param name="cancellationToken">The cancellation token to cancel the request.</param>
+    /// <returns>
+    /// A <see cref="Task"/> representing the <see cref="IApiResponse"/> instance containing the result:
+    /// <list type="table">
+    /// <listheader>
+    /// <term>Status</term>
+    /// <description>Description</description>
+    /// </listheader>
+    /// <item>
+    /// <term>200</term>
+    /// <description>The list of authorized parties for the end user</description>
+    /// </item>
+    /// <item>
+    /// <term>401</term>
+    /// <description>Unauthorized</description>
+    /// </item>
+    /// <item>
+    /// <term>403</term>
+    /// <description>Forbidden</description>
+    /// </item>
+    /// <item>
+    /// <term>503</term>
+    /// <description>Service Unavailable, used when Dialogporten is in maintenance mode</description>
+    /// </item>
+    /// </list>
+    /// </returns>
+    [Headers("Accept: application/json, text/plain")]
+    [Get("/api/v1/enduser/parties")]
+    Task<IApiResponse<Parties>> GetParties(
+        [Property("Altinn.ApiClients.Maskinporten.RequestContext")] MaskinportenRequestContext? requestContext = null, CancellationToken cancellationToken = default);
+}
